(#200) Separate Expression related code into a its own unit
